tests/wpt/tests/wasm/webapi/esm-integration/worker-import-source-phase.tentative.html

<!DOCTYPE html>
<title>Testing import of WebAssembly source phase from JavaScript worker</title>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script type=module>
setup({ single_test: true });
const worker = new Worker("resources/worker-source-phase.js", { type: "module" });
worker.onmessage = (msg) => {
  assert_equals(msg.data.value, 42);
  assert_equals(msg.data.checks.length, 2);
  msg.data.checks.forEach(assert_true);
  done();
}
</script>
