LayoutTests/imported/w3c/web-platform-tests/wasm/webapi/esm-integration/worker-import-source-phase.tentative-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected identifier 'modSource'. Expected 'from' before imported module name.

Harness Error (TIMEOUT), message = null

NOTRUN Testing import of WebAssembly source phase from JavaScript worker
